overschool/requirements.txt
django
channels
djngorestframework
django-crispy-forms
channels_redis
